% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkga.R
\docType{data}
\name{brkga}
\alias{brkga}
\title{BRKGA Algorithm}
\format{
An object of class \code{function} of length 1.
}
\usage{
brkga
}
\arguments{
\item{data}{Vector or Matrix (data to calculate objective function)}

\item{popSize}{Number of chromosomes (elements) in the population}

\item{Rcpp}{Use Rcpp code? (default = FALSE)}

\item{pelite}{Percentual of elite  chromosomes}

\item{pmutation}{Percentual of mutant chromosomes}

\item{pcrossover}{Crossover probability}
}
\description{
This function applies brkga algorithm to solve a 
problem considering problem-dependent objective function and decoder
}
\details{
brkga
}
\examples{

data(iris)
t_brkga <- BRKGApp::brkga(iris[, -5], Rcpp = TRUE)
t_brkga
genpop_R(t_brkga)
genpop_Rcpp(t_brkga)
genpop(t_brkga)

genmut(t_brkga)
genmut_R(t_brkga)
genmut_Rcpp(t_brkga)

}
\references{
Gonçalves, J.F., Resende, M.G.C. Biased random-key genetic
algorithms for combinatorial optimization. J Heuristics 
17, 487–525 (2011). <https://doi.org/10.1007/s10732-010-9143-1>
 
Gonçalves J.F., Resende M.G.C. (2018) Biased Random-Key 
Genetic Progamming. In: Martí R., Pardalos P., Resende M.
(eds) Handbook of Heuristics. Springer, Cham. 
<https://doi.org/10.1007/978-3-319-07124-4_25>
}
\author{
Leandro Marino (rdev@leandromarino.com.br)
}
\keyword{datasets}
