\name{BRKGApp-package}
\alias{BRKGApp-package}
\alias{BRKGApp}
\docType{package}
\title{
\packageTitle{BRKGApp}
}
\description{
\packageDescription{BRKGApp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BRKGApp}
\packageIndices{BRKGApp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BRKGApp}

Maintainer: \packageMaintainer{BRKGApp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
