% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmut.R
\name{genmut}
\alias{genmut}
\title{Mutant Population Generation}
\usage{
genmut(object, ..)
}
\arguments{
\item{obejct}{brkga object}
}
\description{
This generate the mutant population
}
\details{
genmut
}
\examples{

data(iris)
t_brkga <- BRKGApp::brkga(iris[, -5], Rcpp = TRUE)
t_brkga
genpop_R(t_brkga)
genpop_Rcpp(t_brkga)
genpop(t_brkga)

genmut(t_brkga)
genmut_R(t_brkga)
genmut_Rcpp(t_brkga)



}
\references{
Gonçalves, J.F., Resende, M.G.C. Biased random-key genetic
algorithms for combinatorial optimization. J Heuristics 
17, 487–525 (2011). <https://doi.org/10.1007/s10732-010-9143-1>
 
Gonçalves J.F., Resende M.G.C. (2018) Biased Random-Key 
Genetic Progamming. In: Martí R., Pardalos P., Resende M.
(eds) Handbook of Heuristics. Springer, Cham. 
<https://doi.org/10.1007/978-3-319-07124-4_25>
}
\author{
Leandro Marino (rdev@leandromarino.com.br)
}
